refactor: timeDifference 함수를 utils 폴더로 분리
